Changed time datatipe from Long to Double

+++ src/main/java/com/bemba/goalsapi/dto/EntryDto.java
@@ -7,5 +7,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class EntryDto {
 
-	private Long hours;
+	private Double hours;
 }
