+++ src/main/java/com/bemba/goalsapi/entities/Entry.java
@@ -33,7 +33,7 @@ public class Entry implements Serializable {
 	private Long id;
 
 	@Column
-	private Long hours;
+	private Double hours;
 
 	@Column
 	private LocalDate date;
